@@ -1,79 +1,78 @@
 {
-    "name": "bioengine-admin",
-    "version": "3.0.0",
-    "license": "MIT",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/BioWareRu/BioEngine.Admin.git"
-    },
-    "bugs": {
-        "url": "https://github.com/BioWareRu/BioEngine.Admin/issues"
-    },
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "start-hmr": "ng serve --configuration hmr --source-map=false --hmr-warning=false",
-        "start-hmr-sourcemaps": "ng serve --configuration hmr --source-map=true --hmr-warning=false",
-        "build": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --dev",
-        "build-stats": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --dev --stats-json",
-        "build-prod": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod",
-        "build-prod-stats": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod --stats-json",
-        "lint": "ng lint",
-        "bundle-report": "webpack-bundle-analyzer dist/stats.json"
-    },
-    "dependencies": {
-        "@angular/animations": "^7.2.15",
-        "@angular/cdk": "^7.3.7",
-        "@angular/common": "^7.2.15",
-        "@angular/compiler": "^7.2.15",
-        "@angular/core": "^7.2.15",
-        "@angular/flex-layout": "^7.0.0-beta.24",
-        "@angular/forms": "^7.2.15",
-        "@angular/http": "^7.2.15",
-        "@angular/material": "^7.3.7",
-        "@angular/material-moment-adapter": "^7.3.7",
-        "@angular/platform-browser": "^7.2.15",
-        "@angular/platform-browser-dynamic": "^7.2.15",
-        "@angular/router": "^7.2.15",
-        "@ckeditor/ckeditor5-angular": "^1.1.0",
-        "@ckeditor/ckeditor5-build-balloon": "^12.1.0",
-        "@fortawesome/fontawesome-free": "^5.8.2",
-        "angular-oauth2-oidc": "5.0.2",
-        "angular-tree-component": "^8.4.0",
-        "class-transformer": "^0.2.2",
-        "classlist.js": "^1.1.20150312",
-        "core-js": "^2.6.5",
-        "hammerjs": "^2.0.8",
-        "moment": "2.24.0",
-        "ngx-custom-validators": "^7.0.0",
-        "ngx-input-file": "^7.0.4",
-        "ngx-moment": "3.4.0",
-        "object-path": "^0.11.4",
-        "perfect-scrollbar": "^1.4.0",
-        "rxjs": "^6.5.2",
-        "rxjs-compat": "^6.5.2",
-        "url-parse": "^1.4.7",
-        "uuid": "^3.3.2",
-        "web-animations-js": "^2.3.1",
-        "zone.js": "^0.8.29"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^0.13.9",
-        "@angular/cli": "^7.3.9",
-        "@angular/compiler-cli": "^7.2.15",
-        "@angular/language-service": "^7.2.15",
-        "@angularclass/hmr": "^2.1.3",
-        "@types/node": "^12.0.0",
-        "@types/object-path": "0.11.0",
-        "@types/twitter-for-web": "0.0.0",
-        "@types/url-parse": "^1.4.3",
-        "@types/uuid": "^3.4.4",
-        "codelyzer": "^5.0.1",
-        "node-sass": "^4.12.0",
-        "ts-node": "^8.1.0",
-        "tslint": "^5.16.0",
-        "tslint-consistent-codestyle": "1.15.1",
-        "typescript": "^3.2.4",
-        "webpack-bundle-analyzer": "^3.3.2"
-    }
+  "name": "bioengine-admin",
+  "version": "3.0.0",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/BioWareRu/BioEngine.Admin.git"
+  },
+  "bugs": {
+    "url": "https://github.com/BioWareRu/BioEngine.Admin/issues"
+  },
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "start-hmr": "ng serve --configuration hmr --source-map=false --hmr-warning=false",
+    "start-hmr-sourcemaps": "ng serve --configuration hmr --source-map=true --hmr-warning=false",
+    "build": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --dev",
+    "build-stats": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --dev --stats-json",
+    "build-prod": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod",
+    "build-prod-stats": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod --stats-json",
+    "lint": "ng lint",
+    "bundle-report": "webpack-bundle-analyzer dist/stats.json"
+  },
+  "dependencies": {
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/flex-layout": "^7.0.0-beta.24",
+    "@angular/forms": "^8.0.0",
+    "@angular/material": "^7.3.7",
+    "@angular/material-moment-adapter": "^7.3.7",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
+    "@ckeditor/ckeditor5-angular": "^1.1.0",
+    "@ckeditor/ckeditor5-build-balloon": "^12.1.0",
+    "@fortawesome/fontawesome-free": "^5.8.2",
+    "angular-oauth2-oidc": "5.0.2",
+    "angular-tree-component": "^8.4.0",
+    "class-transformer": "^0.2.2",
+    "classlist.js": "^1.1.20150312",
+    "core-js": "^2.6.5",
+    "hammerjs": "^2.0.8",
+    "moment": "2.24.0",
+    "ngx-custom-validators": "^7.0.0",
+    "ngx-input-file": "^7.0.4",
+    "ngx-moment": "3.4.0",
+    "object-path": "^0.11.4",
+    "perfect-scrollbar": "^1.4.0",
+    "rxjs": "^6.5.2",
+    "rxjs-compat": "^6.5.2",
+    "url-parse": "^1.4.7",
+    "uuid": "^3.3.2",
+    "web-animations-js": "github:angular/web-animations-js#release_pr208",
+    "zone.js": "~0.9.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
+    "@angularclass/hmr": "^2.1.3",
+    "@types/node": "^12.0.0",
+    "@types/object-path": "0.11.0",
+    "@types/twitter-for-web": "0.0.0",
+    "@types/url-parse": "^1.4.3",
+    "@types/uuid": "^3.4.4",
+    "codelyzer": "^5.0.1",
+    "node-sass": "^4.12.0",
+    "ts-node": "^8.1.0",
+    "tslint": "^5.16.0",
+    "tslint-consistent-codestyle": "1.15.1",
+    "typescript": "^3.4.5",
+    "webpack-bundle-analyzer": "^3.3.2"
+  }
 }
